db_api/requirements.txt
Django>=5.0,<5.1
django-ninja>=1.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
boto3>=1.35.0
gunicorn==24.1.1
